Redesign network packet classes
This change is inspired by Packet.NET!
This should save a few memory allocations and hopefully also makes the code a little easier to work with.

+++ RyuSocks/SocksClient.cs
@@ -9,7 +9,7 @@ namespace RyuSocks
     {
         public readonly Dictionary<byte, Func<SocksClient, IClientCommand>> Commands = new()
         {
-            { ConnectCommand.Id, (client) => new ConnectCommand(client) },
+            { (byte)ConnectCommand.Id, (client) => new ConnectCommand(client) },
         };
 
         public SocksClient(IPAddress address, int port) : base(address, port)
